specs struct
